Add C code to attempt to run in parallel

+++ algostructures/search/array_search.h
@@ -1,2 +1,2 @@
-extern int search(const int numbers[], const int length, const int key);
-extern int search_parallel(const int numbers[], const int length, const int key);
+extern long long search(const long long numbers[], const long long length, const long long key);
+extern long long search_parallel(const long long numbers[], const long long length, const long long key);
